% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_smu.R
\name{scale_colour_smu}
\alias{scale_colour_smu}
\title{SMU discrete colour scale}
\usage{
scale_colour_smu(...)
}
\arguments{
\item{...}{Arguments passed to ggplot2::scale_colour_manual()}
}
\description{
SMU discrete colour scale
}
