% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_smu.R
\name{scale_fill_smu}
\alias{scale_fill_smu}
\title{SMU discrete fill scale}
\usage{
scale_fill_smu(...)
}
\arguments{
\item{...}{Arguments passed to ggplot2::scale_fill_manual()}
}
\description{
SMU discrete fill scale
}
