% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_smu.R
\docType{data}
\name{smu_colours}
\alias{smu_colours}
\title{SMU colour palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
smu_colours
}
\description{
SMU colour palette
}
\keyword{datasets}
